4.21

+++ controllers/blogs.js
@@ -9,7 +9,7 @@ blogsRouter.get('/', async (req, res) => {
 
   res.json(blogs.map(blog => blog.toJSON()))
 })
-  
+
 blogsRouter.post('/', async (req, res) => {
   const body = req.body
   const decodedToken = jwt.verify(req.token, process.env.SECRET)
@@ -34,8 +34,19 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async(req, res) => {
-  await Blog.findByIdAndRemove(req.params.id)
-  res.status(204).end()
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(req.params.id)
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(req.params.id)
+    return res.status(204).end()
+  }
+  
+  res.status(401).json({ error: 'insufficient rights to delete the blog post' })
 })
 
 blogsRouter.put('/:id', async(req, res) => {
